% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_constant_bamlss}
\alias{gp_constant_bamlss}
\alias{gp_weibull_bamlss}
\alias{gp_sphere_bamlss}
\title{Gaussian process distribution families for bamlss}
\usage{
gp_constant_bamlss(y, D, initial_par, ...)
gp_weibull_bamlss(y, s, Z = NULL, D, q = NULL, initial_par, ...)
gp_sphere_bamlss(y, s, D, initial_par, ...)
}
\arguments{
\item{y}{The GPs to be used as response observations. A list of numeric
vectors.}

\item{D}{The distance matrices of the coordinates. A list of numeric
matrices.}

\item{initial_par}{The initial distributional parameters. A data frame with
one column per distributional parameter.}

\item{...}{Not used.}

\item{s}{The coordinates at which the GPs were observed. A list of numeric
vectors or matrices.}

\item{Z}{Optionally, the design matrices for an additive linear mean
function. A list of numeric matrices.}

\item{q}{Optionally, the scaling factors of the standard deviation of the
GPs at the coordinates. A list of numeric vectors.}
}
\value{
A \code{\link[bamlss:family.bamlss]{family.bamlss}} object.
}
\description{
These functions set up Gaussian process (GP) distribution families for
\link[bamlss:bamlss-package]{bamlss}.
}
\details{
\code{gp_constant_bamlss} uses a constant mean function and a Matérn covariance
function with the smoothness parameter \eqn{\nu} = 1.5. The distributional
parameters of \code{gp_constant_bamlss} are
\itemize{
\item the \strong{mean},
\item the \strong{stddev} (standard deviation),
\item and the \strong{range}.
}

\code{gp_weibull_bamlss} uses a Weibull growth curve as a mean function and a
Matérn covariance function with the smoothness parameter \eqn{\nu} = 1.5.
The distributional parameters of \code{gp_weibull_bamlss} are
\itemize{
\item the \strong{limit} (of the Weibull growth curve),
\item the \strong{shape} (of the Weibull growth curve),
\item the \strong{scale} (of the Weibull growth curve),
\item the \strong{stddev} (standard deviation),
\item and the \strong{range}.
}

Optionally, a linear mean function can be added to the Weibull growth curve
by passing a list of design matrices \code{Z} to \code{gp_weibull_bamlss}. Each column
of the design matrices adds one distributional parameter with the
corresponding column name to the family.

\code{gp_sphere_bamlss} uses a mean function on a sphere and an exponential
covariance function. It can be used to model tree crowns. The distributional
parameters of \code{gp_sphere_bamlss} are
\itemize{
\item the \strong{radius} (of the tree crown),
\item the \strong{south}-bound expansion (of the tree crown),
\item the \strong{height} (of the tree crown),
\item the \strong{stddev} (standard deviation),
\item and the \strong{range}.
}
}
\examples{
\dontrun{
N <- 30
n <- 30

mean <- 0
stddev <- exp(0)
range <- exp(0)

s <- seq.int(0, 1, length.out = n)
D <- as.matrix(dist(s))

mu <- rep.int(mean, n)
Sigma <- stddev^2 * (1 + D / range) * exp(-D / range)
y <- MASS::mvrnorm(N, mu, Sigma)

y <- unlist(apply(y, 1, list), recursive = FALSE)
D <- rep.int(list(D), N)

initial_par <- data.frame(
  mean = rep.int(mean, N),
  stddev = rep.int(stddev, N),
  range = rep.int(range, N)
)

family <- gp_constant_bamlss(y, D, initial_par)
model <- bamlss(rep.int(0, N) ~ 1, family)
joint <- sample_bamlss_gp(model)
}

}
