% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_bamlss_gp}
\alias{sample_bamlss_gp}
\title{Sample from the posterior of a bamlss model with a Gaussian process response}
\usage{
sample_bamlss_gp(model, n = 1000, only_shared = FALSE)
}
\arguments{
\item{model}{A bamlss model with a GP response.}

\item{n}{The sample size/length of the chain to be drawn.}

\item{only_shared}{If \code{TRUE}, only the parameters for those covariates which
appear in the parametric predictors for both the standard
deviation and the range are sampled jointly. Not helpful
in my experience.}
}
\value{
A list with the elements \code{model} and \code{samples}.
}
\description{
This function provides an alternative MCMC sampler for
\link[bamlss:bamlss-package]{bamlss} models with
\link[=gp_constant_bamlss]{Gaussian process (GP) responses}.
It is typically more efficient than the default bamlss sampler by sampling
the parametric predictors for the standard deviation and the range jointly,
thus reducing the autocorrelation of the chains.
}
\examples{
\dontrun{
N <- 30
n <- 30

mean <- 0
stddev <- exp(0)
range <- exp(0)

s <- seq.int(0, 1, length.out = n)
D <- as.matrix(dist(s))

mu <- rep.int(mean, n)
Sigma <- stddev^2 * (1 + D / range) * exp(-D / range)
y <- MASS::mvrnorm(N, mu, Sigma)

y <- unlist(apply(y, 1, list), recursive = FALSE)
D <- rep.int(list(D), N)

initial_par <- data.frame(
  mean = rep.int(mean, N),
  stddev = rep.int(stddev, N),
  range = rep.int(range, N)
)

family <- gp_constant_bamlss(y, D, initial_par)
model <- bamlss(rep.int(0, N) ~ 1, family)
joint <- sample_bamlss_gp(model)
}

}
